+++ rstest_macros/src/parse/arguments.rs
@@ -85,7 +85,7 @@ impl Args {
     fn get(&self, pat: &Pat) -> Option<&ArgumentInfo> {
         self.args
             .get(pat)
-            .or_else(|| self.args.get(&pat_invert_mutability(&pat)))
+            .or_else(|| self.args.get(&pat_invert_mutability(pat)))
     }
 
     fn entry(&mut self, pat: Pat) -> std::collections::hash_map::Entry<Pat, ArgumentInfo> {